back-end/requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-socketio==5.4.1
python-socketio==5.12.0
redis==5.2.1
celery==5.4.0
passlib==1.7.4
bcrypt==4.2.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
eventlet==0.37.0
